+++ assets/release_notes.md
@@ -3,6 +3,7 @@
 
 - Fix action button colors
 - Added Norwegian translations
+- Fix serial number field when creating stock item
 
 ### 0.11.5 - April 2023
 ---